Slit loss correction
We have updated the code to make synthetic spectra from all available
in the NIR (J+H). It seems that the J band is slightly unstable and we
therefore chose to slit loss correct the spectra using the H-band.

When making the VIS+NIR spectra we can include other transmission
curves to make synthetic fluxes in the bluer part and see how the
f_slit_loss evolves as a function of wavelength